+++ sockeye/constants.py
@@ -51,6 +51,11 @@ TARGET_EMBEDDING_PREFIX = "target_" + EMBEDDING_PREFIX
 TARGET_POSITIONAL_EMBEDDING_PREFIX = "target_pos_" + EMBEDDING_PREFIX
 SHARED_EMBEDDING_PREFIX = "source_target_" + EMBEDDING_PREFIX
 
+# source factors
+SOURCE_FACTORS_COMBINE_SUM = 'sum'
+SOURCE_FACTORS_COMBINE_CONCAT = 'concat'
+SOURCE_FACTORS_COMBINE_CHOICES = [SOURCE_FACTORS_COMBINE_SUM, SOURCE_FACTORS_COMBINE_CONCAT]
+
 # encoder names (arguments)
 RNN_NAME = "rnn"
 RNN_WITH_CONV_EMBED_NAME = "rnn-with-conv-embed"
@@ -156,6 +161,17 @@ CNN_ACTIVATION_TYPES = [GLU, RELU, SIGMOID, SOFT_RELU, TANH]
 CNN_PAD_LEFT = "left"
 CNN_PAD_CENTERED = "centered"
 
+# coverage types
+COVERAGE_COUNT = "count"
+COVERAGE_FERTILITY = "fertility"
+COVERAGE_TYPES = [TANH,
+                  SIGMOID,
+                  RELU,
+                  SOFT_RELU,
+                  GRU_TYPE,
+                  COVERAGE_COUNT,
+                  COVERAGE_FERTILITY]
+
 # default I/O variable names
 SOURCE_NAME = "source"
 SOURCE_LENGTH_NAME = "source_length"
@@ -236,12 +252,12 @@ TRAINING_STATE_PARAMS_NAME = "params"
 ARGS_STATE_NAME = "args.yaml"
 
 # Arguments that may differ and still resume training
-ARGS_MAY_DIFFER = ["overwrite_output", "use-tensorboard", "quiet",
+ARGS_MAY_DIFFER = ["overwrite_output", "use_tensorboard", "quiet",
                    "align_plot_prefix", "sure_align_threshold",
                    "keep_last_params", "seed",
-                   "max-updates", "min-updates",
-                   "max-num-epochs", "min-num-epochs",
-                   "max-samples", "min-samples"]
+                   "max_updates", "min_updates",
+                   "max_num_epochs", "min_num_epochs",
+                   "max_samples", "min_samples"]
 
 # Other argument constants
 TRAINING_ARG_SOURCE = "--source"