+++ sockeye/training.py
@@ -420,6 +420,7 @@ class EarlyStoppingTrainer:
     :param model: TrainingModel instance.
     :param optimizer_config: The optimizer configuration.
     :param max_params_files_to_keep: Maximum number of params files to keep in the output folder (last n are kept).
+    :param keep_initializations: Regardless of number of params to keep, never delete the first checkpoint.
     :param source_vocabs: Source vocabulary (and optional source factor vocabularies).
     :param target_vocab: Target vocabulary.
     """
@@ -428,11 +429,13 @@ class EarlyStoppingTrainer:
                  model: TrainingModel,
                  optimizer_config: OptimizerConfig,
                  max_params_files_to_keep: int,
+                 keep_initializations: bool,
                  source_vocabs: List[vocab.Vocab],
                  target_vocab: vocab.Vocab) -> None:
         self.model = model
         self.optimizer_config = optimizer_config
         self.max_params_files_to_keep = max_params_files_to_keep
+        self.keep_initializations = keep_initializations
         self.tflogger = TensorboardLogger(logdir=os.path.join(model.output_dir, C.TENSORBOARD_NAME),
                                           source_vocab=source_vocabs[0],
                                           target_vocab=target_vocab)
@@ -509,6 +512,7 @@ class EarlyStoppingTrainer:
             self._save_params()
             self._update_best_params_link()
             self._save_training_state(train_iter)
+            self._save_initial_optimizer_states(lr_decay_opt_states_reset)
             self._update_best_optimizer_states(lr_decay_opt_states_reset)
             self.tflogger.log_graph(self.model.current_module.symbol)
             logger.info("Training started.")
@@ -758,7 +762,7 @@ class EarlyStoppingTrainer:
         Cleans parameter files, training state directory and waits for remaining decoding processes.
         """
         utils.cleanup_params_files(self.model.output_dir, self.max_params_files_to_keep,
-                                   self.state.checkpoint, self.state.best_checkpoint)
+                                   self.state.checkpoint, self.state.best_checkpoint, self.keep_initializations)
         if process_manager is not None:
             result = process_manager.collect_results()
             if result is not None:
@@ -775,6 +779,10 @@ class EarlyStoppingTrainer:
                 best_opt_states_fname = os.path.join(self.model.output_dir, C.OPT_STATES_BEST)
                 if os.path.exists(best_opt_states_fname):
                     os.remove(best_opt_states_fname)
+            if lr_decay_opt_states_reset == C.LR_DECAY_OPT_STATES_RESET_INITIAL:
+                initial_opt_states_fname = os.path.join(self.model.output_dir, C.OPT_STATES_INITIAL)
+                if os.path.exists(initial_opt_states_fname):
+                    os.remove(initial_opt_states_fname)
 
     def _initialize_parameters(self, params: Optional[str], allow_missing_params: bool):
         self.model.initialize_parameters(self.optimizer_config.initializer, allow_missing_params)
@@ -922,7 +930,7 @@ class EarlyStoppingTrainer:
         """
         self.model.save_params_to_file(self.current_params_fname)
         utils.cleanup_params_files(self.model.output_dir, self.max_params_files_to_keep, self.state.checkpoint,
-                                   self.state.best_checkpoint)
+                                   self.state.best_checkpoint, self.keep_initializations)
 
     def _save_training_state(self, train_iter: data_io.BaseParallelSampleIter):
         """